// lib/schemas.ts
import { z } from "zod";

export const ContactFormSchema = z.object({
    name: z.string().min(1, { message: 'Name is required.' }).min(2, { message: 'Must be at least 2 characters.' }),
    email: z.string().min(1, { message: "Email is required." }).email('Invalid email.'),
    message: z.string().min(1, { message: 'Message is required.' })
})